Move NOOP_CONTROL to storybook utils and change to a standalone file
