+++ packages/grafana-ui/.storybook/preview.ts
@@ -59,5 +59,3 @@ export const parameters = {
     escapeHTML: false,
   },
 };
-
-export const NOOP_CONTROL = { control: { disable: true } };
