+++ packages/grafana-ui/src/components/CallToActionCard/CallToActionCard.story.internal.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { renderComponentWithTheme } from '../../utils/storybook/withTheme';
 import { CallToActionCard, CallToActionCardProps } from './CallToActionCard';
-import { NOOP_CONTROL } from '../../../.storybook/preview';
+import { NOOP_CONTROL } from '../../utils/storybook/noopControl';
 import { Story } from '@storybook/react';
 import { Button } from '../Button/Button';
 import { action } from '@storybook/addon-actions';
