+++ packages/grafana-ui/src/utils/storybook/noopControl.tsx
@@ -1,0 +1,1 @@
+export const NOOP_CONTROL = { control: { disable: true } };
