+++ packages/grafana-ui/tsconfig.build.json
@@ -1,12 +1,4 @@
 {
-  "exclude": [
-    "**/*.story.tsx",
-    "**/*.story.internal.tsx",
-    "**/*.test.ts*",
-    "**/*.tmpl.ts",
-    "dist",
-    "node_modules",
-    "src/utils/storybook"
-  ],
+  "exclude": ["**/*.story.tsx", "**/*.test.ts*", "**/*.tmpl.ts", "dist", "node_modules", "src/utils/storybook"],
   "extends": "./tsconfig.json"
 }
